+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/casGeneral.js
