+++ packages/code-du-travail-frontend/src/outils/indemniteLicenciement/stepStyles.js
